[2.x] refactor: Address review feedback for dependencyMode
- Move DependencyMode enum to sbt.librarymanagement package (lm-core)
  so filtering logic can eventually move there too
- Add DependencyMode type/val to Import.scala for build file access
- Move private filter implementations (filterByDirectDeps, filterByPlusOne,
  isScalaLibraryModule, matchesDirectDep, directDepIndex) from
  Classpaths object to ClasspathImpl object
- Clean up ScalaArtifacts import and remove redundant test task
